Address review comment.
